+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- Added `Late`, `Async`, and `LateAsync`.
+- Results are constructed with factory constructors.
+- Added `Result.guard(fn)`
+- Added `LateResult.fromStream(stream)`
+- Added `AsyncResult.asStream(fn)`
+
 ## 0.1.0
 
 Initial release